Replace original files with refactored files, update import references
