+++ server/facets/api/articleComments.ts
@@ -2,7 +2,7 @@
 import Boom = require('boom');
 import Caching = require('../../lib/Caching');
 import MW = require('../../lib/MediaWiki');
-import Utils = require('../../lib/Utils-refactor');
+import Utils = require('../../lib/Utils');
 import localSettings = require('../../../config/localSettings');
 import wrapResult = require('./presenters/resultWrapper');
