+++ server/facets/api/search.ts
@@ -1,6 +1,6 @@
 import Caching = require('../../lib/Caching');
 import MW = require('../../lib/MediaWiki');
-import Utils = require('../../lib/Utils-refactor');
+import Utils = require('../../lib/Utils');
 import localSettings = require('../../../config/localSettings');
 import wrapResult = require('./presenters/resultWrapper');
