--- server/routes-refactor.ts
@@ -1,83 +1,0 @@
-import Hoek = require('hoek');
-import localSettings = require('../config/localSettings');
-import Caching = require('./lib/Caching');
-
-var unauthenticatedRoutes: any[],
-	unauthenticatedRouteConfig,
-	authenticatedRoutes: any[],
-	indexRoutePaths: string[],
-	proxyRoutePaths: string[];
-
-unauthenticatedRouteConfig = {
-	config: {
-		cache: {
-			privacy: Caching.policyString(Caching.Policy.Public),
-			expiresIn: 60000
-		}
-	}
-};
-
-unauthenticatedRoutes = [
-	{
-		method: 'GET',
-		path: '/favicon.ico',
-		handler: require('./facets/operations/proxyMW')
-	},
-	{
-		method: 'GET',
-		path: '/robots.txt',
-		handler: require('./facets/operations/proxyMW')
-	},
-	{
-		method: 'GET',
-		path: '/front/{path*}',
-		handler: require('./facets/operations/assets')
-	},
-	{
-		method: 'GET',
-		path: '/public/{path*}',
-		handler: require('./facets/operations/assets')
-	},
-	{
-		method: 'GET',
-		path: '/heartbeat',
-		handler: require('./facets/operations/heartbeat')
-	},
-	{
-		method: 'GET',
-		path: localSettings.apiBase + '/article/{articleTitle*}',
-		handler: require('./facets/api/article').get
-	},
-	{
-		method: 'GET',
-		// TODO: if you call to api/v1/comments/ without supplying an id, this actually calls /api/v1/article
-		path: localSettings.apiBase + '/article/comments/{articleId}/{page?}',
-		handler: require('./facets/api/articleComments').get
-	},
-	{
-		method: 'GET',
-		path: localSettings.apiBase + '/search/{query}',
-		handler: require('./facets/api/search').get
-	}
-];
-
-indexRoutePaths = [
-	'/wiki/{title*}',
-	'/{title*}',
-	// TODO this is special case needed for /wiki path, it should be refactored
-	'/{title}'
-];
-
-indexRoutePaths.forEach((path) => {
-	unauthenticatedRoutes.push({
-		method: 'GET',
-		path: path,
-		handler: require('./facets/showArticle')
-	});
-});
-
-unauthenticatedRoutes = unauthenticatedRoutes.map((route) => {
-	return Hoek.applyToDefaults(unauthenticatedRouteConfig, route);
-});
-
-export = unauthenticatedRoutes;
